updates
updated workflows, added example, fix code
